% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g-content.R
\name{g_content}
\alias{g_content}
\title{Retrieve the content portion of a gopher response}
\usage{
g_content(res, has_selector = FALSE)
}
\arguments{
\item{res}{gopher response object}

\item{has_selector}{if \code{TRUE}, only gopher menu entries with populated
selectors (site resource links, URLs, etc) will be returned.
Default: \code{FALSE}}
}
\description{
Retrieve the content portion of a gopher response
}
\examples{
gopher("gopher.floodgap.com/") \%>\%
  g_content()
}
