% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gopher.R
\name{gopher}
\alias{gopher}
\title{Fetch a resource using the gopher protocol}
\usage{
gopher(url, path = NULL, query = NULL, verbose = FALSE)
}
\arguments{
\item{url}{gopher url (with or without the \code{gopher://} prefix)}

\item{path}{optional selector path; this can be part of the \code{url} vs a separate
parameter but this helps avoid having to \code{sprintf()} or \code{paste0()}
values in client code.}

\item{query}{character string containing the search terms/phrase to pass to
a gopher search engine (i.e. a 'veronica' endpoint).}

\item{verbose}{if \code{TRUE} debugging output will be shown; Default: \code{FALSE}}
}
\description{
Fetch a resource using the gopher protocol
}
\examples{
gopher("gopher://gopher.floodgap.com:70/")
gopher("gopher://gopher.floodgap.com/")
gopher("gopher.floodgap.com/") \%>\%
  g_content()
}
