% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\docType{data}
\name{item_type}
\alias{item_type}
\title{Valid single character item types}
\format{An object of class \code{character} of length 23.}
\usage{
item_type
}
\description{
Valid single character item types
}
\examples{
names(item_type)
item_type["I"]
}
\keyword{datasets}
