% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.gopher_response}
\alias{print.gopher_response}
\title{S3 print method for gopher response objects}
\usage{
\method{print}{gopher_response}(x, ...)
}
\arguments{
\item{x}{gopher response object}

\item{...}{unused}
}
\description{
S3 print method for gopher response objects
}
