% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varmint-package.R
\docType{package}
\name{varmint}
\alias{varmint}
\alias{varmint-package}
\title{Retrieve Content from and Interact with 'Gopher' Servers}
\description{
The 'Gopher' protocol is a 'TCP/IP' application layer protocol
designed for distributing, searching, and retrieving documents over the
Internet. The 'Gopher' protocol was strongly oriented towards a menu-document
design and presented an alternative to the World Wide Web in its early
stages, but ultimately 'Hypertext Transfer Protocol' ('HTTP') became the
dominant protocol. The 'Gopher' ecosystem is often regarded as the effective
predecessor of the 'World Wide Web'. Tools are provided to interact with
and retrieve content from 'Gopher' servers.
}
\author{
Bob Rudis (bob@rud.is)
}
